Update output of backup timestamps
GitLab Omnibus command "gitlab-backup restore" and rake task
gitlab:backup:restore restore a previously created GitLab backup.

If multiple GitLab backups exist in the backup directory, it is required
to provide the timestamp of the backup to restore.

If the timestamp of the backup to restore is not given with multiple
backups in the backup directory, the command and rake task print the
timestamps of all backups in the backup directory.

The output of the timestamps is a bit difficult to read:
* The first timestamp is printed prefixed with the current date and
  time.
* Subsequent backup timestamps are printed without the current date and
  time prefix.

This commit updates the output of the second and subsequent backup
timestamps to be prefixed with the current date and time, as it is done
for the first backup timestamp.

Changelog: changed