Support using '|' as a delimiter for search document references
Using a space ' ' as a delimiter is challenging when there are nil
values, and it prevents us from adding more fields to a document
reference string representation.

We are adding deserialization support for both delimiters before
changing our serialization logic to support any instances where
a rollback of GitLab could occur and there would be a mix of
serialization formats.