Update capybara gem to v3.39.0
This update is needed to ensure Ruby 3.1 compatibility.

Ruby 3.1 extracted the `matrix` gem into a separate gem, and as a
result Capybara needs to have the gem explicitly required. This pulls
in https://github.com/teamcapybara/capybara/pull/2468.

Note that the `qa` directory already uses `capybara` v3.39.0.

Now that `Capybara.w3c_click_offset` defaults to `true`, clicks with
x/y coordinates will click in the middle of the element. Setting an
`offset` of 0 will click on the left element.