Simplify parameter lookup in issues_finder_shared_examples.rb
Prior to Ruby 3.1, `WorkItems::WorkItemsFinder::Params` would be
resolved to `IssuesFinder::Params`.

However, with https://github.com/ruby/ruby/pull/4585 to fix
https://bugs.ruby-lang.org/issues/17887, Ruby 3.1 now resolves
`WorkItems::WorksItemsFinder::Params` as `EE::IssuesFinder::Params`.
However, since the latter is a prepended module, it does not have the
constants `FILTER_NONE` or `FILTER_ANY`. As a result, the test fails
with `unintialized constant` errors in Ruby 3.1 and up.

Simplify the test to use `IssuableFinder::Params` since we can use
a straightforward lookup that doesn't depend on the finder.