Sign and verify ansi2json state
This state serves as a sort of a pagination cursor for the jobs
controller trace rendering.

The state is passed in a base64 encoded query parameter, and gets
embedded in the JSON output in a way that makes it look like the values
originate from the GitLab backend itself.

To prevent confusion, we are appending a signature to the state and only
restoring states with a valid signature.

See https://gitlab.com/gitlab-org/gitlab/-/issues/392794

Changelog: fixed