Reschedule migration to add admin_mode scope for PATs
The original background migration failed in certain installations. Token
scopes are serialized by ActiveRecord and stored as strings, but our
documentation contained a script which added a scope as symbol. This led
to the failing jobs.

This commit fixes the migration to consolidate symbol values stored in
scopes when needed, and requeues it.

Changelog: fixed