Skip setting base path for GitLab.com
In the GitLab for Jira app, when we are on GitLab.com
the base path does not need to be set as it will be implied
from the domain that is serving the app. It's only needed
for self-managed when it's manually set by the user.

Not hardcoding this will allow for easier testing on local
and staging environments.

Changelog: changed