Add basic implementation of loading only a single diff file
In `viewDiffsFileByFile` mode, the app will wait until the diffs
metadata is loaded - this basic amount of information is absolutely
crucial to being able to load one file - and then trigger the
`fetchFileByFile` action.

This is INSTEAD of the unscoped triggering of the
`fetchDiffFilesBatch` action, which doesn't need the metadata in
order to begin. `fetchDiffFilesBatch` is entirely skipped in
file-by-file mode.

When using the file-by-file sibling navigation buttons, we also
need to trigger the `fetchFileByFile` action, since simply updating
the current diff file hash in the store is not enough to show that
file (since it's not yet loaded, we have to fetch it).

Changelog: changed