Patch doorkeeper-openid_connect to use RFC7638 thumbprint
We had a production incident in
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8664 when
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116142 updated
`doorkeeper-openid_connect` from v1.8.3 to v1.8.5.

In v1.8.4,
https://github.com/doorkeeper-gem/doorkeeper-openid_connect/pull/177
swapped `json-jwt` in favor of `ruby-jwt`.

The OpenID `kid` is an opaque value that is supposed to identify
public keys. However, `json-jwt` and `ruby-jwt` implement the `kid`
slightly differently: the former generates a thumbprint as described
in RFC7638 (https://www.rfc-editor.org/rfc/rfc7638), while the latter
generates a SHA256 digest based on the key itself.

We could preserve the previous behavior by setting
`JWT.configuration.jwk.kid_generator_type` to `:rfc7638_thumbprint`.
However, this would be a global setting that might affect other users
of the `jwt` gem.

For now, just patch `doorkeeper-openid_connect` to pull in
https://github.com/doorkeeper-gem/doorkeeper-openid_connect/pull/194
to use the RFC7638 type.

Changelog: fixed