Check read permissions in Llm::CompletionWorker
Before we make a call to AI API, we need to check if user who
actually executed the action can read the resource and if the
resource can actually be sent to AI (utilize `#send_to_ai?`).

We already have the permission check at the mutation level and
`#send_to_ai?` check in `Llm::BaseService`. But it is possible
that those permissions changes after the job is enqueued.

No changelog since this is still behind a feature flag.