Return service error when invalid
Currently, we're not returning anything to the user when service
gets executed but it is actually invalid because not all
requirements are met (e.g. resource can't be sent to AI, FF
disabled, etc). We return a success regardless.

This modifies the service layer to return an error if it's
invalid.

No changelog since this is behind a feature flag.