Automatically index projects in Zoekt when namespace is enabled
Zoekt is a new code search database being gradually rolled out on
GitLab.com . It is intended to replace Elasticsearch for code search.

We use the `::Zoekt::IndexedNamespace` model as a way to keep track of
namespaces that are enabled as part of our rollout of using Zoekt for
code search. When we implemented this we did not automate the indexing
of the projects when this was enabled and we had this as another manual
step as part of the rollout. Now that we're building APIs to enable a
namespace in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116650 we want
this to also automatically index all the projects in the namespace when
it is enabled.

This MR just adds an after create hook on the model to trigger a Sidekiq
worker which then loops through all projects in the namespace and
triggers another Sidekiq worker to do the indexing. This is larger a
copy of the way we do it with the `ElasticNamespaceIndexerWorker` and as
such we have added a 2nd argument to the worker which will be used in
the future to handle deletes.

This change also required removing `sidekiq_inline` and tweaking some
specs a little to avoiding the indexing requests being sent to Zoekt
when creating a `::Zoekt::IndexedNamespace`.

Changelog: added
EE: true
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/117634