Handle multiples of the same scanner in one pipeline
This change addresses a bug whereby if multiple instances of the same
scanner run in a single pipeline the last scanner to be ingested will
incorrectly mark all of the vulnerabilities detected by the other
scanner(s) as resolved.

In the current implementation, as each scanner result is ingested the
list of vulnerabilities detected by the scanner are checked against
**all** vulnerabilities detected on the project by the same scanner
type.  Any vulnerabilities found on the project that were not detected
by the scanner are marked as resolved.

So given 'Scanner A' and 'Scanner B' are both the same type, and run
within the same pipeline:

- Scanner A is ingested
- Scanner A detected vulnerability 1
- Detected vulnerabilities are subtracted from project vulnerabilities:
    [1] - [1] = [] => No vulnerabilities to mark as resolved.

- Scanner B is ingested
- Scanner B detected vulnerability 2
- Detected vulnerabilities are subtracted from project vulnerabilities:
    [1, 2] - [2] = [1] => Vulnerability 1 is incorrectly marked as
                          resolved.

By pulling the 'mark as resolved' step up a level the detected
vulnerabilities can all be grouped by scanner type and the incorrect
resolution will no longer be marked.

- Scanner A vulnerabilities + Scanner B vulnerabilities = [1, 2]
- Project vulnerabilities for scanner type = [1, 2]
    [1, 2] - [1, 2] = [] => No vulnerabilities to mark as resolved.