Fix Ci::CommitWithPipeline batch loader key
There was a bug in a `BatchLoader` call within `Ci::CommitWithPipeline`
where cross-project calls could cause incorrect results. This bug had no
customer impact because the conditions required to trigger it were never
met.

The bug was that the loader block referred to the instance's `project`
attribute, so when called for multiple projects prior to the loader's
evaluation, the loader would only load data for one of the projects.

The fix is to tell the `BatchLoader` to include `project.id` in the
batch key.

Changelog: fixed