Fix merge request Jira sync problems due to branch
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/115448 added a
change that sends through branch data to Jira alongside the merge
request data.

It introduced 2 bugs, one when the flag was disabled, and one when it
was enabled.

This line
https://gitlab.com/gitlab-org/gitlab/-/blob/44461cd6b2ed95ad70e147d4c880486effa5ce2b/app/workers/jira_connect/sync_merge_request_worker.rb#L21
where we add the associated branch to the sync needed to be wrapped in
the feature flag too.

When the flag was disabled, we were sending Jira a branch that had no
Jira reference refs (unless the branch specifically had the Jira issue
reference in its name), and Jira then would reject the whole request.

When the flag is enabled, the branch does include Jira issue refs.

When enabled, there was a bug where closing a merge request would not
update the merge request's status in Jira. This was because we only
select open merge requests to send through within the branch Jira issue
references, and if the merge request was closed, this would lead to the
request being rejected by Jira for similar reasons as above.

This change:

- Sends nil branches when
  jira_include_keys_from_associated_mr_for_branch flag is disabled
- Sends nil branches when the merge request is not open

https://gitlab.com/gitlab-org/gitlab/-/issues/354373

Changelog: fixed