Trigger deploys to release-environments on stable branches
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2696

This change expands upon the `start-release-environments` child pipeline
to do the following

* Correct how we tag images built in `cng-mirror` by leveraging the
`IMAGE_TAG_EXT` field to uniquely tag images. Some examples of this
working are at

https://gitlab.com/gitlab-org/build/CNG-mirror/-/jobs/4090399906 15-9-stable

https://gitlab.com/gitlab-org/build/CNG-mirror/-/jobs/4090296623 15-10-stable

* Adds a new stage to actually do a downstream pipeline trigger to the
`release-environments` repository in order to actually trigger a deploy.
This passes in the environment to deploy to and a json object which is
the versions of each component to deploy. The work for the
release-environments repository to support this was done in
https://gitlab.com/gitlab-com/gl-infra/release-environments/-/merge_requests/51