Add support for snowplow PA configurator
Since product analytics is moving from
using Jitsu to Snowplow to ingest event data
we need to change the onboarding setup to connect
to snowplow, not jitsu.

This commit adds a new application
setting to store the root URI for the
configurator as well as a new project setting
to store the instrumentation key receiveed
from the configurator service.

All of this new functionality is behind a
default-off feature flag and needs
to be turned on to test and use.

Changelog: added