Dry up whether super sidebar should be shown
This moves a few instances of `current_user.use_new_navigation` to use
`NavHelper.show_super_sidebar?` instead. This has the benefit, that we
can now change the logic in a single place (e.g. once the feature flag
becomes the default).

Furthermore we are able to remove the `super_sidebar_supported?`
private method and it's use, because _all_ sidebars support the super
sidebar since:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/115162 is merged.