Create SnapshotBuilder
SnapshotBuilder is a helper class that
accepts a sorted list of resource events of the following types
for a timebox:

- ResourceTimeboxEvent
- ResourceStateEvent
- ResourceWeightEvent
- ResourceLinkEvent

SnapshotBuilder summarizes the end states of
of the all issues and work items for each day in the timebox range.

For example, if "Issue A" is assigned to "Timebox 1" and given
a weight of 1 then closes the next day, we would have two daily
snapshots for the issue based on all resource events for the issue.

In pseudocode, SnapshotBuilder would return something like:

Snapshots = [
  {
    Date: Today - 1,
    "Issue A": { timebox: "Timebox 1", weight: 1, state: :open }
  }
  {
    Date: Today,
    "Issue A": { timebox: "Timebox 1", weight: 1, state: :closed }
  }
]