Execute OpenAI request via Workhorse
The send-url injector is now flexible enough to perform any
external request. We can use it for our experimental AI API
to make it more robust.

1. User performs a request to Workhorse
2. Workhorse performs a request to Rails
3. Rails responds with send-url header and sends all the
necessary information: auth header, params
4. Workhorse recognizes the response as a send-url injector
5. Workhorse performs the API request and sends the response to
the user

Problems that are fixed:

- Workhorse can handle too long and too many requests
- Less load to Gitlab Rails

Features added:

- Streaming is now supported

Security question:

- The auth headers are sent to Workhorse now. It should be
fine because they are not visible to the external viewer.