Move merge base pipeline logic to EE
`::Ci::CompareMetricsReportsService`,
`::Ci::CompareCodequalityReportsService`, and
`::Ci::CompareSecurityReportsService` are all EE classes. Since
`use_merge_base_pipeline_for_comparison?` is related to those classes
specifically, it should be in EE. This change moves the code to the EE
file. To lower the risk of EE class usage happening the in the CE file
in the future, we use constants instead of strings for the map keys.
This means that trying to reference an EE constant inside CE will cause
an undefined constant error and fail the pipeline when running tests.