Make assets image build on stable branches
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2674

Currently we do a trigger to the `cng-mirror` repository to build images
on every commit to a stable branch, because we want to deploy the images
built into a stable environment (via trigger to `release-environments).

However, the rules on when we build the assets image (which we need) vs
when we trigger `cng-mirror` and `release-environments` are different.

This causes a failure when we haven't got the assets image tag available
to use to build images and thus a subsequent deploy also fails (normally
job dependencies would stop the deploy due to images failing to build,
but as these jobs are in `allow_failure: true` this doesn't currently
happen.

This change adds the specific rule about building on stable branches
(that we use to trigger the whole release-environments child pipeline)
to the rules for the `build-assets-image` job. This means whenever
we would deploy to a release-environment, we are guarenteed to have
run the `build-assets-image` job.