Remove job_age from jobs/request API endpoint
This feature was likely never working as intended because of caching,
and the parameter does not seem to be in use on GitLab.com.

The removal is behind a feature flag that is _enabled by default_, but
will be explicitly disabled on GitLab.com prior to merging this MR. That
way, if we by any chance miss the feature flag cleanup in 16.0, the
feature will still be removed.

This removal was announced in 15.2, see
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/91509

See https://gitlab.com/gitlab-org/gitlab/-/issues/334253

Changelog: removed