Delete work items global id route
The work items global id route is deprecated and is now ready
for deletion in milestone 16.0.

There is a lot of logic in the frontend code for checking whether
to use the work item id or iid, so instead of deleting all this
logic (which results in a huge MR diff) `fetchByIid` is set to
`true`. Subsequent MRs will clean up and delete the rest of the
frontend logic.