Ensure project is persisted when generating new issue link
This ensures that a project actually exists before attempting to
generate the "Create issue" shortcut link for it. This fixes an issue
where, if you visited the "Create project" page as an admin with the new
nav enabled, the page would error-out. The reason seems to be that, on
that page, a `Project` is instantiated, but not persisted, so the left
side of the condition passed, bringing us to the permission check. For
some reason, `can?` returned `true`, indicating that an admin is allowed
to create issues in a project that does not exist. This might be the
intneded behavior though, and we should definitely check whether the
project is persisted first.