Add includes and construct values necessary for unit test
It seems all prior testing assumed a different - integration -
test context that had many of these classes and values already
constructed. When attempting to unit test the `diffs_tab_pane_data`
method with a stand-alone unit test, it failed. These new includes
and constructing a current_user allow it to be tested without the
context of a full session.