Fix incompatibility of gopsutil dependency with macOS
We are seeing some compiler warnings when compiling workhorse on macOS

The actual bug is in the gopsutil package, which forgot to include
the <TargetConditionals.h> header. This was fixed in gopsutil v3.21.3,
but we are still pulling in v3.21.2.

Fix this bug by upgrading to the most recent bug release of the v3.21
release train, which is v3.21.12.

See also: https://gitlab.com/gitlab-org/gitaly/-/merge_requests/5570