Allow overriding of protected ref access level const
Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/12776

We need to allow an additional access level for EE, to do so we need to
override the human access levels. To facilitate this change we must
first make the constant into a class method so we can override it.

No user facing changes.