Fix flaky and quarantined project mirror specs
The tests were failing intermittently because they would input the
URL and submit the data faster than the JavaScript was able to update
the internal field. This caused a blank URL to be submitted on
occasion. To make this specs more reliable, ensure the hidden URL
field is updated with the right value before clicking the "Save
changes" button.

This failed in
https://gitlab.com/gitlab-org/gitlab/-/jobs/4164769751. A similar
problem was fixed in other merge requests:

* https://gitlab.com/gitlab-org/gitlab/-/merge_requests/92178
* https://gitlab.com/gitlab-org/gitlab/-/merge_requests/95778

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/408519