Introduced chat capabilities to Explain Code
Minor refactoring for constants, etc.

Fine-tuned styling for the AiGenieChat component

Updated styling of the code blocks in AI responses

Unified Explain Code to use chat interface

Now. no matter whether we request chat or not,
we route even the very first request to the chat
functionality and processing

Refactored AiGenieChat component to match drawer

Removed bottom margin from the last node

When in aiGenieChat message, we do not want the
unexpected bottom margins to affect the height of
the message

Added the top bordre to AiGenieChat

Removed avatars from the chat

Moved chat error closer to input

Slightly updated the tokens computation

Reset messages and error on new explaiantion

When requesting explanation of  a new code
snippet, we reset the current messages and the
error message to start blank with the new snippet

Simplified messages prop in chat component

Abstracted all API into the utility

Moved the selected text into a slot

This allows to generalize the chat component more

Simplified the loading flag

Removed unnecessary asyncs

Improved the soft scrolling on the new messages

Re-named the method to generalize

Simplified utilities for easier replacement

Updated styles for the chat component

Extended the tests coverage of AiGenie component

Removed style changes to minimze the scope

Tested the utilities

Updated tests for AiGenie component