DRY up security configuration unassign
Use Security::Orchestration::UnassignService to unassign security
orchestration configurations in AssignService.

In case a project should get a new security configuration assigned
but it already has one, we need to unassign the old configuration first.
There is already a service to do this. This commit makes use of the
service in order to remove repeated code and make the unassign
process consistent.