Allow using db replicas for GraphQL subs
We were always querying the primary when executing
database queries from GraphQL subscription queries
in order to avoid replication lag.

This change uses a similar strategy to what we do
for Sidekiq: pass the WAL through redis so that
we can make smarter decisions at the receiving
end for which host to use.

Changelog: performance