Sign ansi2json state by default
This enables the flag sign_and_verify_ansi2json_state by default. The
flag should in principle also be safe to remove, but that's a slightly
bigger code change that has a greater chance of back-and-forth with
reviewers. I have an upcoming holiday and want to make sure this change
makes it into 16.0 in some form so we can close out the associated
issue.

See https://gitlab.com/gitlab-org/gitlab/-/issues/404718

Changelog: changed