Fix expanded avatars list display in the MR widget
We recently changed how avatars render in the MR widget by switching
from a block to an inline layout. This broke the layout when many users
have approved an MR. In this case, the list can be expanded to show the
full list, which doesn't play well with an inline layout as the list
overflows the widget instead of wrapping as it should.

Changelog: fixed