Merge all Verify CODEOWNERS sections
By merging all Verify related sections under a single [Verify]
section we ensure that a single set of required approvals is
applied.

For example: if a CI template is modified we apply the specific
approval rule which overrides the catch-all rules at the top of
Verify section.

We want the Verify section to be a single channel of approvals
that has generic catch-all rules but also specific rules.