Reimplement peek behaviour
This reimplements the sidebar peek behaviour that was originally added
in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116914.

This implementation uses a document-wide `mousemove` listener, whereas
the previous implementation relied on the `mouseenter/leave` events
fired on the sidebar element itself and a small hover area.

The old approach
----------------

The main problem with the previous implementation was that sometimes the
sidebar would oscillate between peeked open and closed if the cursor
was stationary.

This was sometimes easy to reproduce, and sometimes hard. The reason
comes down to a race between the JavaScript timer and the CSS transition
duration.

The [UI Events specification][spec] says of the `mouseenter` event:

[spec]: https://w3c.github.io/uievents/#event-type-mouseenter

> A user agent MUST also dispatch this event when the element or one of
> its descendants moves to be underneath the primary pointing device.

It turns out that browsers do not obey this as you might expect on
elements that are undergoing transitions under a stationary cursor (I
suspect this applies to animations as well). Presumably this is for
performance reasons.

What they actually do in this case is dispatch the event approximately
*after* the transition has ended. Each browser does this slightly
differently in my testing, and none of them seem deterministic in
exactly when the event is dispatched.

This means that, if the sidebar's opening transition is cancelled before
it is complete (e.g., the sidebar instead starts its closing transition)
the `mouseenter` does not get dispatched even if the transition did move
the element underneath the stationary cursor.

If the cursor is moving at the same time as the transitioning element,
the event is dispatched as expected on the first frame the cursor enters
it.

So, the race was between the opening sidebar transition and the closing
timer. If the transition ended early enough, the `mouseenter` event
would be dispatched, and the bug wouldn't happen. If the cursor moved
during the transition and was over the transitioning sidebar, the
`mouseenter` event would be dispatched, as expected.

The bug occurred only when the cursor was stationary, and the opening
transition took too long, causing it to be cancelled by the closing
transition, meaning no `mouseenter` event was dispatched to keep the
sidebar open.

The new approach
----------------

The approach taken here is to rely on a document-wide `mousemove`
listener. This has two benefits:

- The bug with the previous approach is avoided since no DOM is involved
  in determining where the cursor is.
- Additional regions are trivial to add, for instance, moving the cursor
  far enough away instantly closes the sidebar, rather than having to
  wait for the closing timer. It is possible to other regions, for
  instance, one which shows a hint or affordance of the peek behaviour,
  but does not start the timer to open the sidebar (this is not
  implemented here, but is trivial to do so).

It does have one drawback, which is that it's not coupled to the DOM.
This means that if the sidebar changes width, or has some descendant
that overflows the sidebar, the sidebar might get closed even though the
cursor is over it. This could be avoided in a few ways, but isn't done
here.

This also formalises and extracts the peek state machine logic into its
own Vue component. This means that the document event listener is
guaranteed to be added and cleaned up when switching between toggled
open/closed states. The timers are also encapsulated within that
component, and the whole thing can be more thoroughly tested in
isolation.