Add `ref_path` to CI job JWTs
Add `ref_path` with the fully qualified ref to CI job JWTs in order to
avoid ambiguity. Branches and tags might have the same name, which
would result in a collision when trying to use only the `ref` field.
For example, if both a tag and branch exist named `v0.0.1`, we'd have
these fields:

- Branch: `{ ref: 'v0.0.1', ref_path: 'refs/heads/v0.0.1' }`
- Tag: `{ ref: 'v0.0.1', ref_path: 'refs/tags/v0.0.1' }`

Changelog: added