Fix for the rebase merge request state being shown incorrectly
Previously, the following was happening:

1. Merge request form gets submitted.
2. User gets direct to merge request page, which loads the widget.
3. The value for `ff_merge_possible?` gets cached incorrectly.

In between steps 1 and 2, the async diff creation is happening, which
actually fetches the ref into the project. If the ref is fetched after
the page is loaded, the widget will cache the wrong value, and the
state won't expire until either the target branch SHA changes or the
diff head SHA changes.

Changelog: fixed