doc/blueprints: Add blueprint for iterating on object pool design
The current design of object pools in Gitaly has organically grown and
is showing limits of what it can do. In order to be able to iterate on
its implementation details, we in the Gitaly team want to take ownership
of the lifecycle of object pools so that we've go tighter control of how
exactly they work. This gives us the ability to iterate faster and
change their inner workings to be more scalable in the future without
impacting callers.

Add an architecture blueprint that describes the problem, the proposed
solution as well as alternative solutions we could implement.