Make it possible to run Puma v6
Puma v6.0 dropped `nakayoshi_fork` in
https://github.com/puma/puma/issues/2925 since "it has very minimal
gain and can trigger nasty bugs in c-extensions due to usage of
GC.compact." If the call to `nakayoshi_fork` is made with Puma 6, the
Web server will crash and fail to start up.

To avoid this, conditionally enable it only if Puma v5 is loaded.

Changelog: fixed