Create bot user when security policy is linked
This creates a security_policy_bot user and adds it as a member to
a project when a new security policy project is linked.

The bot user will be user to run scan execution pipelines in the next
iteration.

The change is added behind the scan_execution_bot_users feature flag.

Changelog: added