Fix some CODEOWNERS rules that missed some approvers
With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/118461 we
consolidated the CODEOWNERS file but misunderstood how default section
code owners work.

If a line-specific code owner is specified, the default section code
owner isn't used at all. If we want to add the line-specific code owners
to the default section one, we need to repeat it with the line-specific
code owners.

Signed-off-by: Rémy Coutable <remy@rymai.me>