Make DesignManagement::Repository a container for GitRepository
This commit improves the re-factoring introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/111555,
such that DesignManagement::Repository is now a container for
DesignManagement::GitRepository.

We no longer need to delegate missing methods from
DesignManagement::Repository to DesignManagement::GitRepository.

It also makes it so that `project.design_repository` is a
`DesignManagement::GitRepository` which is of the kind `Repository`
 again.

However, `project.design_management_repository` is a
`DesignManagement::Repository`.

Changelog: other