Switch the single-file file-by-file feature flag to default to true
This enables the feature that will only load a single file when
using the MR "View one file at a time" checkbox.

Previously, using this checkbox would still load every file in the
MR, which could lead to long delays before being able to view the
file - depending on the situation - and did lead to unnecessarily
loading files that weren't currently being reviewed.