Enable "inherit_approval_rules_on_creation" feature flag
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/408218

**Problem**

We don't create approval rules for merge request unless they are
explicitly provided by the user.

It currently works for creating MR via UI, because frontend provides a
list of approval rules in `approval_rules_attributes` param.

It doesn't work for merge_request API endpoint, because we don't expose
of populate this attribute.

**Solution**

The code is complicated enough and requires a refactoring. The least
invacive approach is to generate `approval_rules_attributes` list if it
was not provided.

That should ensure that MergeRequest::CreateService will inherit
project approval rules even if the user didn't provide them.

Changelog: fixed