Fix pages unique domain redirect
For root domain urls, like `namespace.gitlab.io`, the unique pages
redirect was not working.

The root cause was that the `prefix` was wrongly built for unique domain
enabled projects since the `Project#pages_url` was returning the unique
domain url, making its comparison with the root domain url never match.

To fix that, the `Project#pages_url` now accepts a argument to force the
unique domain, keeping the original behavior on most places and only
returning the unique url when it's needed.

Related to: https://gitlab.com/gitlab-org/gitlab-pages/-/issues/1058
Changelog: fixed