Fix log format string quotation in nginx config
If the log format is not enclosed in 'single quotes' the spaces inside
the log format are ignored (syntactical whitespace). The "double quotes"
are then interpreted, meaning the spaces inside "double quotes" work but
the quotes themselves are not actually output to the log files.

See also the "log_format combined" predefinition in nginx docs.
https://nginx.org/en/docs/http/ngx_http_log_module.html#log_format