Introduce dedicated instrumentation storage
This is a light abstraction over SafeRequestStore
that uses a subset of its interface, but stores
all values written in a dedicated bucket.

This allows us to easily throw out instrumentation
data when instrumenting more than one unit of work.
This is necessary, for example, when dealing with
websocket pushes instead of HTTP requests.

Not a user facing change.