Add helpful 400 message when enabling Zoekt indexing when FF is disabled
This was raised in https://gitlab.com/gitlab-org/gitlab/-/issues/407331
. If a user tries to enabling indexing for a namespace or explicitly
trigger indexing for a project they will just get back 200 from this API
but nothing will happen. This may be confusing for self-managed admins
setting this up and we could save them some time debugging if we return
a 400 early and just say that it can't be done.

We still want to allow the other APIs in here when the feature flag is
disabled because they may still reasonably want to disable a specific
namespace or figure out what is being indexed before they turn a feature
flag back on.