Make SSO button larger
We increase the relevance and noticeability of the SSO buttons by making
them the same width as the normal GitLab `Sign in` button. As decided
here:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/116995#note_1358733815

In addition, when the `restyle_login_page` feature flag is enabled, we
remove the label `Sign in` and replaced by a horizontal separator. This
is inline with the overall minimalistic design of the restyle. In
addition, it makes it more similar to the trial registration page.

Changelog: changed

EE: true