Add a migration to add prefix project in wiki rid
This MR will add the prefix project in rid in all wiki_blobs. This is
needed because in the future we will support group wikis as well so we
need a way to differentiate the rid in project wikis and group wikis.

Changelog: other
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/119261
EE: true