GH-style Jira OAuth endpoints default return 404
The endpoints in `Oauth::JiraDvcs::AuthorizationsController` have been
deprecated since 15.1.

Due to uncertainty about the impact of a full removal in 16.0, all
endpoints now return `404` by default but we allow customers to toggle a
flag to reverse this breaking change.

At time of writing the flag is enabled for GitLab.com and will be
removed closed to 16.0 release to make a breaking change for SaaS
customers.

See:

https://gitlab.com/gitlab-org/gitlab/-/issues/362168#note_1359618964
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/118126#note_1375655863

In future we will make the breaking change irreversible
https://gitlab.com/gitlab-org/gitlab/-/issues/408148.

Changelog: removed