Cleanup artefacts from notes.id bigint migration
For couple of tables, `merge_request_user_mentions` and
`issue_user_mentions`, where we had to create index and FK in
advance, we have incorrect timestamps for migrations. This was OK for
GitLab.com, as MRs were merged and deployed in the correct order, but
could be an issue for development and test environments.

`issue_user_mentions` we fixed with
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/117476. This does
the same for `merge_request_user_mentions`, and in addition

* adds some more cleanup in the migrations tests
* fixes a typo
* adds additional cleanup migration to drop the indexes and FKs if they
  exists. This could happen for development environments only.

https://gitlab.com/gitlab-org/gitlab/-/issues/392815

Changelog: added