Prevent reuse of credit card that was previously used by a banned user
Update the credit card verification step of identity verification to add
a check ensuring the credit card provided by the user has not been
previously used by a banned user. If the provided credit card details
are reused the step fails and the user is prompted to try a different
credit card.

The check is done by searching for Users::CreditCardValidation records
associated with a Users::BannedUser that matches the following details
of the provided credit card:

Card expiration date
Last four digits of the card
Card network (Visa, Mastercard, etc.)
Cardholder name